OnTouchListener implemented
OnClick as shown in Monday's demonstration only allows a single change when the button is clicked the first time.

I looked into how to make the program more responsive and implemented OnTouchListener which allows for "held down"/"release" actions. Holding the button down changes the text/button color to green and makes the text larger, while releasing returns the elements back to normal.

Also commented the code a bit